+++ data/raw_data/extract_raw_data.R
@@ -1,11 +1,16 @@
 # load packages
 library(tidyverse)
+library(googlesheets4)
+
+# site corrections
+sites <- read_sheet("https://docs.google.com/spreadsheets/d/1fq2owxMBLBwwibcdw2uQQFxnIhsMbaH4Qcj_xUwVvSQ/edit?usp=sharing", 2) %>% #nolint
+  separate_rows(site_raw, sep = ",")
 
 # define were the raw inventories data are
-path <- "../inventories/data/raw_data/all_output_files/"
+path <- "../inventories/data/raw_data"
 
 # check available inventories
-files <- list.files(path, full.names = TRUE, pattern = ".csv")
+files <- list.files(path, full.names = TRUE, pattern = "harmonized_data")
 files
 
 # explore taxonomy extraction
@@ -17,8 +22,10 @@ read_csv(file, locale = readr::locale(encoding = "latin1")) %>%
   ))) %>%
   # in case a column doesn't exist but this should not be the case ?
   unique() %>%
-  mutate_all(as.character)
-# in case a column is badly read
+  mutate_all(as.character) %>%
+  # in case a column is badly read
+  rename(site_raw = Site) %>%
+  left_join(sites)
 
 # make it a function
 extract_taxo <- function(file) {
@@ -28,7 +35,9 @@ extract_taxo <- function(file) {
       "Family", "Genus", "Species"
     ))) %>%
     unique() %>%
-    mutate_all(as.character)
+    mutate_all(as.character) %>%
+    rename(site_raw = Site) %>%
+    left_join(sites)
 }
 
 # test
@@ -38,5 +47,11 @@ extract_taxo(files[1])
 raw_taxo <- lapply(files, extract_taxo) %>%
   bind_rows()
 
+# check that no site are missing
+raw_taxo %>%
+  filter(is.na(site)) %>%
+  select(site_raw, site) %>%
+  unique()
+
 # save
-write_tsv(raw_taxo, "data/raw_data/raw_taxonomy_v1.tsv")
+write_tsv(raw_taxo, "data/raw_data/raw_taxonomy_v3.tsv")