+++ NEWS.md
@@ -1,3 +1,4 @@
+-   species 0.1.0.9019: data v3 with 25 sites
 -   species 0.1.0.9018: try v2 #2
 -   species 0.1.0.9017: data and outputs description
 -   species 0.1.0.9016: v2 without the TRY request